look ma... no useful commit messages
I guess I got to a part where this is useful to both myself and someone
else somewhere

a range is essentially a predicate, so no need to have an extra class
for it

sometimes I got different precisions for versions I'm parsing, so it's
useful to have a way of enforcing them and make sure things can work
as intended without having to mess around the internal representation

and now I'm asking myself: is this dialog box a chat? is it a place for
me to dump my thoughts?

congrats to you reading this... because it might be the last piece of
written commit message not generated by an AI

nah!!! I'm kidding